+++ react/components/TopMenu.js
@@ -2,20 +2,24 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { intlShape, injectIntl } from 'react-intl'
 import SearchBar from 'vtex.storecomponents/SearchBar'
+import Logo from 'vtex.store-components/Logo'
 
 import { ExtensionPoint } from 'render'
 
-const TopMenu = ({ name, intl, fixed }) => {
+const TopMenu = ({ logoUrl, logoTitle, intl, fixed }) => {
   const translate = id => intl.formatMessage({ id: `dreamstore.${id}` })
 
   return (
-    <div className={`${fixed ? 'fixed mt4 top-2 shadow-5' : 'top-0'} z-999 flex flex-wrap w-100 pa4 pa5-ns ph7-l bg-white tl`}>
+    <div className={`${fixed ? 'fixed mt4 top-2 shadow-5' : 'top-0'} z-999 flex flex-wrap justify-center w-100 pa4 pa5-ns ph7-l bg-white tl`}>
       <div className="flex pa4">
-        <a className="link b f3 near-black tc tl-ns serious-black" href="/">
-          {name}
+        <a className="link b f3 near-black tc tl-ns serious-black flex" href="/">
+          <Logo 
+            url={logoUrl}
+            title={logoTitle}
+          />
         </a>
       </div>
-      <div className="flex items-center flex-auto">
+      <div className="vtex-searchbar__container flex items-center flex-auto">
         <div className="w-100 flex pr8-ns">
           <div className="w-100">
             <SearchBar
@@ -33,7 +37,8 @@ const TopMenu = ({ name, intl, fixed }) => {
 }
 
 TopMenu.propTypes = {
-  name: PropTypes.string.isRequired,
+  logoUrl: PropTypes.string.isRequired,
+  logoTitle: PropTypes.string.isRequired,
   intl: intlShape.isRequired,
   fixed: PropTypes.bool,
 }