+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Integration of logo component.
+
 ## [1.0.0] - 2018-6-4
 ### Added
 - Add the breadcrumb component to the SearchPage and ProductPage